no local history based on live flights
i realised that the live flights were stored in hypertables with unique timestamps which is not handy to analyze since anyways i'm not gonna analyze live session flights but fetched historic data from opensky, so now live flights are in normal tables and the unique key is icao24 which the code takes care of not alias
SUMMARY: for a given live flight during session, there is only one line in the database.
NEXT STEP: we will be able to use a live flight (ex: clicked-on) so fetch its entire history from internet and analyze only THAT history
MAIN TAKEAWAY: the data we generate during the open session is NOT MEANINGFUL for a typical flight with hours of history (minutes against hours is nothing) so we just don't care saving this small history anymore!