Flights legacy table & all dependencies cleanup
The  table was meant to be used later for historical data, but even then, involving an external Python script would only add extra runtime latency. Also, updating the C++/Qt/Python pipelines has an extra cost we can easily avoid for now.

Historical data can later be stored in separate table, triggered by the analysis parts in the UI, on demand and only for specific flights (in terms of computational and storage constraint, it doesn't make sense to pull historical data for an entire fetched file with tens and hundreds of real time flights)